dry out trigger classes
