+++ denorm/db/postgresql/triggers.py
@@ -132,63 +132,21 @@ CREATE TRIGGER %(name)s
 
 
 class TriggerSet(base.TriggerSet):
-    def drop_atomic(self):
+    def cursor(self):
+        if not hasattr(self, '_cursor'):
+            cursor = super(TriggerSet, self).cursor()
+            cursor.execute("SELECT lanname FROM pg_catalog.pg_language WHERE lanname ='plpgsql'")
+            if not cursor.fetchall():
+                cursor.execute('CREATE LANGUAGE plpgsql')
+            self._cursor = cursor
+        return self._cursor
+
+    def drop_trigger(self, trigger_name, table_name):
         qn = self.connection.ops.quote_name
         cursor = self.cursor()
-        for trigger_name, table_name in self.installed_triggers():
-            cursor.execute('DROP TRIGGER %s ON %s;' % (qn(trigger_name), qn(table_name)))
-
-    def drop_unneeded_atomic(self):
-        qn = self.connection.ops.quote_name
-        cursor = self.cursor()
-        needed_triggers = self.triggers.keys()
-        for trigger_name, table_name in self.installed_triggers():
-            if trigger_name in needed_triggers:
-                continue
-            print 'Dropping trigger on %s' % table_name
-            cursor.execute('DROP TRIGGER %s ON %s;' % (qn(trigger_name), qn(table_name)))
-
-    def drop_unneeded(self):
-        try:
-            with transaction.atomic():
-                self.drop_unneeded_atomic()
-        except AttributeError:
-            self.drop_unneeded_atomic()
-            transaction.commit_unless_managed(using=self.using)
+        cursor.execute('DROP TRIGGER %s ON %s;' % (qn(trigger_name), qn(table_name)))
 
     def installed_triggers(self):
         cursor = self.cursor()
         cursor.execute("SELECT pg_trigger.tgname, pg_class.relname FROM pg_trigger LEFT JOIN pg_class ON (pg_trigger.tgrelid = pg_class.oid) WHERE pg_trigger.tgname LIKE 'denorm_%%';")
         return cursor.fetchall()
-
-    def drop(self):
-        try:
-            with transaction.atomic():
-                self.drop_atomic()
-        except AttributeError:
-            self.drop_atomic()
-            transaction.commit_unless_managed(using=self.using)
-
-    def install_atomic(self):
-        cursor = self.cursor()
-        cursor.execute("SELECT lanname FROM pg_catalog.pg_language WHERE lanname ='plpgsql'")
-        if not cursor.fetchall():
-            cursor.execute('CREATE LANGUAGE plpgsql')
-        installed_triggers = self.installed_triggers()
-        if len(installed_triggers) > 0:
-            installed_triggers = zip(*installed_triggers)[0]  # get just the names
-        for name, trigger in self.triggers.items():
-            if name in installed_triggers:
-                continue
-            sql, args = trigger.sql()
-            cursor.execute(sql, args)
-            yield trigger
-
-    def install(self):
-        try:
-            with transaction.atomic():
-                ret = list(self.install_atomic())
-        except AttributeError:
-            ret = list(self.install_atomic())
-            transaction.commit_unless_managed(using=self.using)
-        return ret
